Add slice offsets to the Milty Draft for other player counts
This adds the offsets for the hexes that make up each slice when not playing with exactly six players.
`OFFSETS_HYPER` has the offsets for when the standard ring of 83A-88A hyperlanes are to the left of a player, when there are 5 or fewer players.
`OFFSETS_7C_8CG` has the offsets for some players in the seven and eight player maps. If the bottom player in the rulebook map images is player "A", and the other players are labelled clockwise, then these offsets are for player C in seven and eight player games, as well as player G in eight player games. These players have the 83B/85B hyperlane tiles directly in front of their home system.
`OFFSETS_7E` has the offsets for player E in the seven player map following the labelling above. They have the 88B hyperlane tile where their equidistant would be.
This change does not implement these offsets into the tool, only provides the positions for a later change to implement.